Refactor: 데이터를 외부에서 관리하여 CalendarScreen 과 ScheduleScreen에서 쉽게 사용할 수 있도록 코드 구조 변경.
