+++ app/src/main/java/com/erica/gamsung/uploadTime/presentation/CalendarView.kt
@@ -21,7 +21,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateMapOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
@@ -47,6 +46,7 @@ const val CHUNK_NUM = 7
 @Composable
 fun CalendarView(
     // focusedDate: LocalDate? = null,
+    selectedDatesMap: Map<YearMonth, List<LocalDate>>,
     onDateSelected: ((LocalDate, Boolean) -> Unit)? = null,
     onToggleValid: Boolean,
 ) {
@@ -54,11 +54,10 @@ fun CalendarView(
         mutableStateOf(YearMonth.now())
     }
     val now = YearMonth.now()
-    val selectedDatesMap =
-        remember {
-            mutableStateMapOf<YearMonth, List<LocalDate>>()
-        }
-
+//    val selectedDatesMap =
+//        remember {
+//            mutableStateMapOf<YearMonth, List<LocalDate>>()
+//        }
     var moveDirection by remember {
         mutableStateOf(CalendarMoveType.START)
     }
@@ -78,17 +77,11 @@ fun CalendarView(
         }
     }
 
-    fun toggleDateSelection(
+    fun toggleDateSelectionInUI(
         date: LocalDate,
         isSelected: Boolean,
     ) {
-        val currentSelectedDates = selectedDatesMap[currentYearMonth]?.toMutableList() ?: mutableListOf()
-        if (isSelected) {
-            currentSelectedDates.remove(date)
-        } else {
-            currentSelectedDates.add(date)
-        }
-        selectedDatesMap[currentYearMonth] = currentSelectedDates
+        // UI에서 날짜 선택 토글을 처리하는 대신, 상위 컴포넌트에게 변경 사항을 알립니다.
         onDateSelected?.invoke(date, !isSelected)
     }
 
@@ -129,7 +122,7 @@ fun CalendarView(
         CalendarGrid(
             yearMonth = currentYearMonth,
             selectedDatesMap = selectedDatesMap,
-            onDateSelected = { date, isSelected -> if (onToggleValid) toggleDateSelection(date, isSelected) },
+            onDateSelected = { date, isSelected -> if (onToggleValid) toggleDateSelectionInUI(date, isSelected) },
         )
     }
 }
@@ -165,7 +158,7 @@ fun DaysOfWeekRow() {
     val daysOfWeek =
         listOf(DayOfWeek.SUNDAY) +
             DayOfWeek
-                .values()
+                .entries
                 .filter { it != DayOfWeek.SUNDAY }
     Row {
         daysOfWeek
@@ -225,16 +218,15 @@ fun CalendarGrid(
             week
                 .forEach { day ->
                     if (day > 0 && day != FILTER_NUM) {
+                        val date = yearMonth.atDay(day)
+                        val isSelected = selectedDatesMap[yearMonth]?.contains(date) ?: false
                         DateView(
                             modifier =
                                 Modifier
                                     .weight(1f),
-                            date =
-                                yearMonth
-                                    .atDay(day),
-                            currentYearMonth = yearMonth,
-                            selectedDatesMap = selectedDatesMap,
-                            onDateSelected = { selectedDate, isSelected ->
+                            date = date,
+                            isSelected = isSelected,
+                            onDateSelected = { selectedDate, _ ->
                                 // 선택된 날짜를 처리 하는 로직
                                 onDateSelected?.invoke(selectedDate, !isSelected)
                                 // 변경된 날짜 목록을 상위 Component 로 전달
@@ -262,19 +254,10 @@ fun CalendarGrid(
 fun DateView(
     modifier: Modifier = Modifier,
     date: LocalDate,
-    currentYearMonth: YearMonth,
     focusedDate: LocalDate? = null,
-    selectedDatesMap: Map<YearMonth, List<LocalDate>>,
+    isSelected: Boolean,
     onDateSelected: ((LocalDate, Boolean) -> Unit)? = null,
 ) {
-    // UI의 즉각적인 변화를 위해 selectedDAtesMap[currentYearMonth] 직접 참조
-    val isSelected =
-        remember(
-            selectedDatesMap[currentYearMonth],
-            date,
-        ) {
-            selectedDatesMap[currentYearMonth]?.contains(date) ?: false
-        }
     val isFocused = date == focusedDate
     val backGroundColor =
         when {
